Refactor common/models.py
* Move out all classes in common/models that are not a Page or a
  Snippet.
* Make the block models small files that are imported into
  blocks/models.py then can be imported into other apps.
