+++ blocks/__init__.py
@@ -1,0 +1,21 @@
+COLOUR_CHOICES = [
+    ('white', 'White'),
+    ('grey', 'Grey'),
+    ('blue', 'Blue'),
+]
+
+COLUMN_CHOICES = [
+    ('12', '12/12'),
+    ('11', '11/12'),
+    ('10', '10/12'),
+    ('9', '9/12'),
+    ('8', '8/12'),
+    ('7', '7/12'),
+    ('6', '6/12'),
+    ('5', '5/12'),
+    ('4', '4/12'),
+    ('3', '3/12'),
+    ('2', '2/12'),
+    ('1', '1/12'),
+    ('0', '0/12'),
+]
