+++ blocks/models.py
@@ -1,0 +1,12 @@
+from .centered_text import CenteredTextBlock
+from .custom_image import ImageCustomBlock
+from .google_map import GoogleMapBlock
+from .hero import HeroBlock
+from .job import JobsBlock
+from .people import PeopleBlock
+from .photo_stream import COSPhotoStreamBlock
+from .spotlight import SpotlightBlock
+from .tab import TabIndexBlock, TabContainerBlockInColumn, TabBlock, TabContainerBlock
+from .twitter import TwitterBlock
+from .two_column import TwoColumnBlock
+from .three_column import ThreeColumnBlock
