+++ blocks/people.py
@@ -1,0 +1,17 @@
+from wagtail.wagtailcore import blocks
+
+PEOPLE_DISPLAY_CHOICES = [
+    ('concise-team', 'concise-team'),
+    ('concise-ambassador', 'concise-ambassador'),
+    ('detailed', 'detailed'),
+]
+
+
+class PeopleBlock(blocks.StructBlock):
+    displayStyle = blocks.ChoiceBlock(choices=PEOPLE_DISPLAY_CHOICES,default="concise")
+    tag = blocks.CharBlock(max_length=20)
+
+    class Meta:
+        template = 'common/blocks/people_block.html'
+        icon = 'group'
+        label = "PeopleBlock"
