+++ common/models.py
@@ -1,4 +1,3 @@
-
 from django.db import models
 from django.shortcuts import render
 
@@ -27,378 +26,17 @@ from taggit.models import TaggedItemBase
 from taggit.managers import TaggableManager
 from modelcluster.contrib.taggit import ClusterTaggableManager
 
-
-COLOUR_CHOICES = [
-    ('white', 'White'),
-    ('grey', 'Grey'),
-    ('blue', 'Blue'),
-]
-
-COLUMN_CHOICES = [
-    ('12', '12/12'),
-    ('11', '11/12'),
-    ('10', '10/12'),
-    ('9', '9/12'),
-    ('8', '8/12'),
-    ('7', '7/12'),
-    ('6', '6/12'),
-    ('5', '5/12'),
-    ('4', '4/12'),
-    ('3', '3/12'),
-    ('2', '2/12'),
-    ('1', '1/12'),
-    ('0', '0/12'),
-]
-
-IMAGE_STYLE_CHOICES = [
-    ('max-width:225px;max-height:145px', 'small display'),
-    ('max_width:250px;max-height:250px', 'middle display'),
-    ('max_width:250px;max-height:250px;padding-top:20px', 'middle + padding display'),
-    ('height:auto', 'auto display'),
-]
-
-PEOPLE_DISPLAY_CHOICES = [
-    ('concise-team', 'concise-team'),
-    ('concise-ambassador', 'concise-ambassador'),
-    ('detailed', 'detailed'),
-]
-
-
-class ClearfixBlock(blocks.StructBlock):
-    class Meta:
-        template = 'common/blocks/clearfix_block.html'
-        icon = 'placeholder'
-        label = 'Clearfix Block'
-        help_text = ('When you need to make sure that your next element(s) is on a new line from '
-                     'the previous elements, use this little helper block.')
-
-
-class EmbedBlock(blocks.StructBlock):
-    url = blocks.CharBlock(required=True, max_length=255)
-    height = blocks.IntegerBlock(default=500)
-    width = blocks.IntegerBlock(default=750)
-
-    class Meta:
-        template = 'common/blocks/embed_block.html'
-        icon = 'image'
-        label = 'Embed Youtube Video'
-
-class SpotlightBubbleBlock(blocks.StructBlock):
-    image = ImageChooserBlock()
-    title = blocks.CharBlock(required=True, max_length=35)
-    description = blocks.RichTextBlock(required=True)
-
-    class Meta:
-        template = 'common/blocks/spotlight_bubble_block.html'
-        icon = 'image'
-        label = 'Spotlight Bubble Block'
-
-class SpotlightBlock(blocks.StructBlock):
-    bubbles = blocks.StreamBlock([
-        ('bubble_block', SpotlightBubbleBlock()),
-    ])
-
-    class Meta:
-        template = 'common/blocks/spotlight_block.html'
-        icon = 'image'
-        label = 'Spotlight Block'
-
-
-class HeroBlock(blocks.StructBlock):
-    image = ImageChooserBlock(required=True)
-    description = blocks.RichTextBlock(required=True)
-
-    class Meta:
-        template = 'common/blocks/hero_block.html'
-        icon = 'image'
-        label = 'Hero Block'
-
-
-class GoogleMapBlock(blocks.StructBlock):
-    address = blocks.CharBlock(required=True,max_length=255)
-    map_zoom_level = blocks.CharBlock(default=14,required=True,max_length=3)
-
-    class Meta:
-        template = 'common/blocks/google_map.html'
-        icon = 'cogs'
-        label = 'Google Map'
-
-
-class COSPhotoStreamBlock(blocks.StructBlock):
-
-    class Meta:
-        template = 'common/blocks/flickr.html'
-        icon = 'image'
-        label = 'Photo Stream'
-
-
-class TwitterBlock(blocks.StructBlock):
-    username = blocks.CharBlock(required=True)
-
-    class Meta:
-        template = 'common/blocks/twitter.html'
-        icon = 'placeholder'
-        label = 'Twitter Stream'
-
-
-class ImageCustomBlock(blocks.StructBlock):
-    main_image = ImageChooserBlock()
-    style = blocks.ChoiceBlock(choices=IMAGE_STYLE_CHOICES,default="height:auto")
-    url = blocks.CharBlock(max_length=250, required=False)
-
-    class Meta:
-        template = 'common/blocks/image_custom_block.html'
-        icon = 'image'
-        label = 'Customed Image'
-
-
-class PeopleBlock(blocks.StructBlock):
-    displayStyle = blocks.ChoiceBlock(choices=PEOPLE_DISPLAY_CHOICES,default="concise")
-    tag = blocks.CharBlock(max_length=20)
-
-    class Meta:
-        template = 'common/blocks/people_block.html'
-        icon = 'group'
-        label = "PeopleBlock"
-
-
-class CenteredTextBlock(blocks.StructBlock):
-    text = blocks.RichTextBlock()
-
-    class Meta:
-        template = 'common/blocks/centered_text.html'
-        icon = 'openquote'
-        label = 'Centered Text Block'
-
-
-class ThreeColumnBlock(blocks.StructBlock):
- 
-    background = blocks.ChoiceBlock(choices=COLOUR_CHOICES,default="white")
-    left_column = blocks.StreamBlock([
-            ('heading', blocks.CharBlock(classname="full title")),
-            ('paragraph', blocks.RichTextBlock()),
-            ('image', ImageChooserBlock(template='common/blocks/image.html')),
-            ('appeal', blocks.StructBlock([
-                    ('icon', blocks.ChoiceBlock(required=True, choices=[
-                        ('none', 'none'),
-                        ('flask', 'flask'),
-                        ('group', 'group'),
-                        ('laptop', 'laptop'),
-                        ('sitemap', 'sitemap'),
-                        ('user', 'user'),
-                        ('book', 'book'),
-                        ('download', 'download'),
-                    ])),
-                    ('topic', blocks.CharBlock(required=True, max_length=35)),
-                    ('content', blocks.TextBlock(required=True, max_length=255)),
-            ], classname='appeal', icon='tick', template='common/blocks/appeal.html')),
-            ('embedded_video', EmbedBlock()),
-            ('google_map', GoogleMapBlock()),
-            ('twitter_feed', TwitterBlock()),
-            ('photo_stream', COSPhotoStreamBlock()),
-            ('centered_text', CenteredTextBlock()),
-            ('clear_fixblock', ClearfixBlock()),
-            ('embed_block', EmbedBlock()),
-        ], icon='arrow-left', label='Left column content', classname='col4')
- 
-    center_column = blocks.StreamBlock([
-            ('heading', blocks.CharBlock(classname="full title")),
-            ('paragraph', blocks.RichTextBlock()),
-            ('image', ImageChooserBlock(template='common/blocks/image.html')),
-            ('appeal', blocks.StructBlock([
-                    ('icon', blocks.ChoiceBlock(required=True, choices=[
-                        ('none', 'none'),
-                        ('flask', 'flask'),
-                        ('group', 'group'),
-                        ('laptop', 'laptop'),
-                        ('sitemap', 'sitemap'),
-                        ('user', 'user'),
-                        ('book', 'book'),
-                        ('download', 'download'),
-                    ])),
-                    ('topic', blocks.CharBlock(required=True, max_length=35)),
-                    ('content', blocks.TextBlock(required=True, max_length=255)),
-            ], classname='appeal', icon='tick', template='common/blocks/appeal.html')),
-            ('embedded_video', EmbedBlock()),
-            ('google_map', GoogleMapBlock()),
-            ('twitter_feed', TwitterBlock()),
-            ('photo_stream', COSPhotoStreamBlock()),
-            ('centered_text', CenteredTextBlock()),
-            ('embed_block', EmbedBlock()),
-            ('clear_fixblock', ClearfixBlock()),
-        ], icon='arrow-right', label='Center column content', classname='col4')
-    
-    right_column = blocks.StreamBlock([
-            ('heading', blocks.CharBlock(classname="full title")),
-            ('paragraph', blocks.RichTextBlock()),
-            ('image', ImageChooserBlock()),
-            ('appeal', blocks.StructBlock([
-                    ('icon', blocks.ChoiceBlock(required=True, choices=[
-                        ('none', 'none'),
-                        ('flask', 'flask'),
-                        ('group', 'group'),
-                        ('laptop', 'laptop'),
-                        ('sitemap', 'sitemap'),
-                        ('user', 'user'),
-                        ('book', 'book'),
-                        ('download', 'download'),
-                    ])),
-                    ('topic', blocks.CharBlock(required=True, max_length=35)),
-                    ('content', blocks.TextBlock(required=True, max_length=255)),
-            ], classname='appeal', icon='tick', template='common/blocks/appeal.html')),
-            ('embedded_video', EmbedBlock()),
-            ('google_map', GoogleMapBlock()),
-            ('twitter_feed', TwitterBlock()),
-            ('photo_stream', COSPhotoStreamBlock()),
-            ('centered_text', CenteredTextBlock()),
-            ('embed_block', EmbedBlock()),
-            ('clear_fixblock', ClearfixBlock()),
-        ], icon='arrow-right', label='Right column content', classname='col4')
- 
-    class Meta:
-        template = 'common/blocks/three_column_block.html'
-        icon = 'placeholder'
-        label = 'Three Columns'
-
-
-class TabBlockInColumn(blocks.StructBlock):
-    id = blocks.CharBlock(required=True)
-    isActive = blocks.BooleanBlock(default=False, required=False)
-    container = blocks.StreamBlock([
-        ('paragraph', blocks.RichTextBlock()),
-        ('people_block', PeopleBlock()),
-        ('raw_html', blocks.RawHTMLBlock(
-            help_text='With great power comes great responsibility. This HTML is unescaped. Be careful!')),
-
-    ])
-
-    class Meta:
-        template = 'common/blocks/tab_block.html'
-        icon = 'plus'
-        label = 'Tab'
-
-
-class TabContainerBlockInColumn(blocks.StructBlock):
-    tabs = blocks.StreamBlock([('tab', TabBlockInColumn())])
-    class Meta:
-        template = 'common/blocks/tabs_container_block.html'
-        icon = 'placeholder'
-        label = 'Tab Container'
-
-
-class TabIndexEntryBlock(blocks.StructBlock):
-    id = blocks.TextBlock(max_length=25, required=True)
-    display = blocks.TextBlock(max_length=40, required=True)
-
-    class Meta:
-        icon = 'arrow-right'
-        label = 'Tab Entry'
-
-
-class TabIndexBlock(blocks.StructBlock):
-    display_style = blocks.ChoiceBlock(required=True, choices=[
-                        ('vertical', 'vertical'),
-                        ('horizontal', 'horizontal')])
-    tabsIndexes = blocks.StreamBlock([('tab', TabIndexEntryBlock()),
-                                      ])
-
-    class Meta:
-        template = 'common/blocks/tab_index_block.html'
-        icon = 'list-ul'
-        label = "Tab Indexing"
-
-
-class TwoColumnBlock(blocks.StructBlock):
-    
-    background = blocks.ChoiceBlock(choices=COLOUR_CHOICES,default="white")
-    left_column_size = blocks.ChoiceBlock(choices=COLUMN_CHOICES,default="6")
-    right_column_size = blocks.ChoiceBlock(choices=COLUMN_CHOICES, default="6")
-    left_column = blocks.StreamBlock([
-            ('heading', blocks.CharBlock(classname="full title")),
-            ('paragraph', blocks.RichTextBlock()),
-            ('customized_image', ImageCustomBlock()),
-            ('appeal', blocks.StructBlock([
-                    ('icon', blocks.ChoiceBlock(required=True, choices=[
-                        ('none', 'none'),
-                        ('flask', 'flask'),
-                        ('group', 'group'),
-                        ('laptop', 'laptop'),
-                        ('sitemap', 'sitemap'),
-                        ('user', 'user'),
-                        ('book', 'book'),
-                        ('download', 'download'),
-                    ])),
-                    ('topic', blocks.CharBlock(required=True, max_length=35)),
-                    ('content', blocks.TextBlock(required=True, max_length=255)),
-            ], classname='appeal', icon='tick', template='common/blocks/appeal.html')),
-            ('tab_index', TabIndexBlock()),
-            ('centered_text', CenteredTextBlock()),
-            ('embed_block', EmbedBlock()),
-            ('clear_fixblock', ClearfixBlock()),
-            ('raw_html', blocks.RawHTMLBlock(
-                    help_text='With great power comes great responsibility. This HTML is unescaped. Be careful!')),
-        ], icon='arrow-left', label='Left column content', classname='col4')
- 
-    right_column = blocks.StreamBlock([
-            ('heading', blocks.CharBlock(classname="full title")),
-            ('paragraph', blocks.RichTextBlock()),
-            ('customized_image', ImageCustomBlock()),
-            ('appeal', blocks.StructBlock([
-                    ('icon', blocks.ChoiceBlock(required=True, choices=[
-                        ('none', 'none'),
-                        ('flask', 'flask'),
-                        ('group', 'group'),
-                        ('laptop', 'laptop'),
-                        ('sitemap', 'sitemap'),
-                        ('user', 'user'),
-                        ('book', 'book'),
-                        ('download', 'download'),
-                    ])),
-                    ('topic', blocks.CharBlock(required=True, max_length=35)),
-                    ('content', blocks.TextBlock(required=True, max_length=255)),
-            ], classname='appeal', icon='tick', template='common/blocks/appeal.html')),
-            ('embedded_video', EmbedBlock()),
-            ('google_map', GoogleMapBlock()),
-            ('twitter_feed', TwitterBlock()),
-            ('embedded_tab_container', TabContainerBlockInColumn()),
-            ('centered_text', CenteredTextBlock()),
-            ('embed_block', EmbedBlock()),
-            ('clear_fixblock', ClearfixBlock()),
-            ('raw_html', blocks.RawHTMLBlock(
-                help_text='With great power comes great responsibility. This HTML is unescaped. Be careful!')),
-        ], icon='arrow-right', label='Right column content', classname='col4')
- 
-    class Meta:
-        template = 'common/blocks/two_column_block.html'
-        icon = 'placeholder'
-        label = 'Two Columns'
-
-
-class TabBlock(blocks.StructBlock):
-    id = blocks.CharBlock(required=True)
-    isActive = blocks.BooleanBlock(default=False, required=False)
-    container = blocks.StreamBlock([
-        ('two_column_block', TwoColumnBlock()),
-        ('paragraph', blocks.RichTextBlock()),
-        ('people_block', PeopleBlock()),
-    ])
-    class Meta:
-        template = 'common/blocks/tab_block.html'
-        icon = 'plus'
-        label = 'Tab'
-
-
-class TabContainerBlock(blocks.StructBlock):
-    tabs = blocks.StreamBlock([('tab', TabBlock())])
-    class Meta:
-        template = 'common/blocks/tabs_container_block.html'
-        icon = 'placeholder'
-        label = 'Tab Container'
+from blocks.models import (
+    ImageCustomBlock, GoogleMapBlock, HeroBlock, JobsBlock, PeopleBlock,
+    COSPhotoStreamBlock, SpotlightBlock, TwitterBlock, CenteredTextBlock,
+    TabIndexBlock, TabContainerBlock, TabContainerBlockInColumn, TwoColumnBlock,
+    ThreeColumnBlock
+)
 
 
 @register_snippet
 class Person(ClusterableModel, index.Indexed):
-    
+
     first_name = models.CharField(max_length=255)
     middle_name = models.CharField(max_length=255, null=True, blank=True)
     last_name = models.CharField(max_length=255)
@@ -464,7 +102,7 @@ class PersonTag(TaggedItemBase):
 @register_snippet
 class Footer(models.Model):
 
-    title = models.CharField(default='untitled', max_length=255)    
+    title = models.CharField(default='untitled', max_length=255)
 
     active = models.BooleanField(default=False)
 
@@ -499,42 +137,8 @@ class Footer(models.Model):
             for footer in Footer.objects.filter(active=True):
                 footer.active = False
                 footer.save()
-            
+
         return super(Footer, self).save(*args, **kwargs)
-
-
-
-class JobBlurb(blocks.StructBlock):
-    title = blocks.CharBlock(required=True, max_length=255)
-    description = blocks.RichTextBlock()
-
-    class Meta:
-        template = 'common/blocks/job_blurb.html'
-        icon = 'placeholder'
-        label = 'Job Blurb Box'
-
-
-class JobListing(blocks.StructBlock):
-    job_title = blocks.CharBlock(required=True, max_length=255)
-    description = blocks.StreamBlock([
-        ('blurb', JobBlurb()),
-    ])
-
-    class Meta:
-        template = 'common/blocks/job_listing.html'
-        icon = 'placeholder'
-        label = 'Job listing'
-
-
-class JobsBlock(blocks.StructBlock):
-    jobs = blocks.StreamBlock([
-        ('job_listing', JobListing()),
-    ])
-
-    class Meta:
-        template = 'common/blocks/jobs_block.html'
-        icon = 'placeholder'
-        label = 'Jobs Block'
 
 
 class HomePage(Page):
